Move setSessionCookie to sessions/http/
