+++ packages/core/src/sessions/http/set-session-cookie.ts
@@ -1,5 +1,5 @@
 // FoalTS
-import { Config, CookieOptions, HttpResponse } from '../core';
+import { Config, CookieOptions, HttpResponse } from '../../core';
 import {
   SESSION_DEFAULT_COOKIE_HTTP_ONLY,
   SESSION_DEFAULT_COOKIE_NAME,
@@ -7,8 +7,8 @@ import {
   SESSION_DEFAULT_CSRF_COOKIE_NAME,
   SESSION_DEFAULT_SAME_SITE_ON_CSRF_ENABLED,
   SESSION_USER_COOKIE_NAME,
-} from './constants';
-import { Session } from './session';
+} from '../constants';
+import { Session } from '../session';
 
 /**
  * Sends the session token in a cookie.
